updated with separate models
