+++ models/food.js
@@ -1,11 +1,15 @@
-module.exports = function(sequelize, DataTypes) {
-  var food = sequelize.define("food", {
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    } //,
-    // createdAt: Sequelize.DATE,
-    // updatedAt: Sequelize.DATE
+// Sequelize (capital) references the standard library
+var Sequelize = require("sequelize");
+// sequelize (lowercase) references my connection to the DB.
+var sequelize = require("../config/connection.js");
+
+// Creates a model that matches up with DB
+var Food = sequelize.define("food", {
+    food: Sequelize.STRING
   });
-  return food;
-};
+  
+// Syncs with DB
+  Food.sync();
+
+  // Makes the Food Model available for other files (will also create a table)
+  module.exports = Food;
