+++ models/symptom.js
